+++ src/aigc_content_studio/__init__.py
@@ -11,6 +11,7 @@ from .workflow import ContentProductionWorkflow
 from .review_decisions import build_human_review_export
 from .review_history import validate_review_history
 from .feedback_replay import replay_reviewer_feedback
+from .review_visibility import summarize_stale_feedback
 
 __all__ = [
     "AssetLedger",
@@ -33,5 +34,6 @@ __all__ = [
     "build_human_review_export",
     "validate_review_history",
     "replay_reviewer_feedback",
+    "summarize_stale_feedback",
 ]
 __version__ = "1.0.0"